+++ test/check_rbtree.cpp
@@ -18,7 +18,7 @@
 
 #include "check_util.h"
 
-void insertProbeIter(int  NUM_ENTRIES)
+void insertProbeIter(size_t NUM_ENTRIES)
 {
 
     //data generation
@@ -41,7 +41,7 @@ void insertProbeIter(int  NUM_ENTRIES)
     std::set<datatuple, datatuple> rbtree;
     int64_t datasize = 0;
     std::vector<pageid_t> dsp;
-    for(int i = 0; i < NUM_ENTRIES; i++)
+    for(size_t i = 0; i < NUM_ENTRIES; i++)
     {
         //prepare the key
         datatuple newtuple;        
@@ -50,7 +50,7 @@ void insertProbeIter(int  NUM_ENTRIES)
         *newtuple.keylen = keylen;
         
         newtuple.key = (datatuple::key_t) malloc(keylen);
-        for(int j=0; j<keylen-1; j++)
+        for(size_t j=0; j<keylen-1; j++)
             newtuple.key[j] = key_arr[i][j];
         newtuple.key[keylen-1]='\0';
 
@@ -60,7 +60,7 @@ void insertProbeIter(int  NUM_ENTRIES)
         *newtuple.datalen = datalen;
         
         newtuple.data = (datatuple::data_t) malloc(datalen);
-        for(int j=0; j<datalen-1; j++)
+        for(size_t j=0; j<datalen-1; j++)
             newtuple.data[j] = data_arr[i][j];
         newtuple.data[datalen-1]='\0';
 
@@ -81,7 +81,7 @@ void insertProbeIter(int  NUM_ENTRIES)
 
     printf("\nTREE STRUCTURE\n");
     //ltable.get_tree_c1()->print_tree(xid);
-    printf("datasize: %d\n", datasize);
+    printf("datasize: %lld\n", (long long)datasize);
 
     printf("Stage 2: Looking up %d keys:\n", NUM_ENTRIES);
 
@@ -93,7 +93,7 @@ void insertProbeIter(int  NUM_ENTRIES)
         //get the key
         uint32_t keylen = key_arr[ri].length()+1;        
         datatuple::key_t rkey = (datatuple::key_t) malloc(keylen);
-        for(int j=0; j<keylen-1; j++)
+        for(size_t j=0; j<keylen-1; j++)
             rkey[j] = key_arr[ri][j];
         rkey[keylen-1]='\0';
